Move program to its own mod
